ADD mailcatcher
[Mailcatcher](https://github.com/sj26/mailcatcher) is a smtp proxy that
catch any outbound mail and presents it in a web interface, ensuring no
email goes outside our dev box.

This will make it easier to debug mail related features, which are
currently about sending mails to real people or not sending mails at
all.

Project owner recommands not to include it in Gemfile, so it has to be
installed locally by developpers :

    gem install mailcatcher

Then, to be able to see mails running :

    mailcatcher

this will create a server on port 1080 (specified in command output).

It is ok not to install mailcatcher or running it as it won't hurt
application runtime - you just won't be able to see mails.

Accordingly, code that enables / disables mail sending has been removed,
since getting mails content is just about running mailcatcher.

Close #12

Details :

* ADD mailcatcher configuration
* ADD doc about mailcatcher in readme
* REMOVE dev mail sending toggler