ADD integration test layer
A lot of errors happen while running the app in browser and are simply
not catched in tests.

That's because only unit testing is performed, giving a false sense of
security.

We need to add an integration test layer as well to be sure features
actually work.

As this has not been done from start and writing a complete test suite
would take too long, let just add new integration tests for bugs we
resolve and features we add.

Needed for #13